+++ NEWS.md
@@ -1,5 +1,32 @@
 # esqlabsR (development version)
 
+## Breaking changes
+
+- Protein ontogenies for individuals and populations are now defined in one column
+`Protein Ontogenies` in the sheets `IndividualBiometrics` (for individuals) or `Demographics`
+(for populations). The columns `Protein` and `Ontogeny` are no longer supported. The new column
+`Protein Ontogenies` is a comma-separated list of protein names and ontogeny names pairs. For example: 
+`CYP3A4:CYP3A4,CYP2D6:CYP2C8` will create a CYP3A4 ontogeny for the protein CYP3A4 and a CYP2D6 ontogeny for the protein CYP2C8. (#825)
+
+## Breaking changes
+
+- The function `createDataCombinedFromExcel()` gets a new signature. The arguments 
+`file` and `sheet` are removed. The file from which the `DataCombined` objects 
+are created is now passed as part of the `ProjectConfiguration` passed as 
+`projectConfiguration` argument, the sheet is always `DataCombined`.
+
+- The function `createDataCombinedFromExcel()` gets a new argument `plotGridNames`.
+The `plotGridNames` argument is a character vector of names of the plots 
+specified in the sheet `plotGrids`. The function will then create and return 
+`DataCombined` used in the specified plots. The new argument can be combined with 
+`dataCombinedNames`.Useful in combination with the new argument `dataCombinedList` 
+of the function `createPlotsFromExcel()`.
+
+- Argument `dataCombinedNames = NULL` of the function `createDataCombinedFromExcel()`
+does not create `DataCombined` for all entries in the excel file any more. If 
+`dataCombinedNames = NULL`, `plotGridNames` must be specified. If both arguments
+are `NULL`, an empty list is returned.
+
 ## Major changes
 
 - User-defined parameters passed to the `createScenarios()` or `Scenario$new()` 
@@ -8,17 +35,26 @@ were overwritten by the administration protocol (\#817).
 - Added `snapshotProjectConfiguration()` and `restoreProjectConfiguration()` functions 
 to support version control and project sharing of project configurations.
 
+- `createPlotsFromExcel()` now accepts a (named) list of `DataCombined` objects as input
+  to create plots defined in the `plotGridNames` argument. Missing `DataCombined`
+  will be created from the Excel file (default behavior).
+
 ## Minor improvements and bug fixes
 
 - `readScenarioConfigurationFromExcel()` ignores rows where `Scenario_name` is empty.
 - Fixed a bug when the dimension in the y-axis label of `sensitivityTimeProfiles()` 
 did not match the unit (\#823).
+- `sensitivityTimeProfiles()` accepts a `DataSet` or a list of `DataSet` objects for `observedData` (\#831).
+- When creating a scenario, the name of the scenario is set as the name of the simulation.
+This way, when saving the simulation to PKML and loading in MoBi, the loaded simulation
+will have the updated name.
 
 # esqlabsR 5.4.0
 
 ## Breaking changes
 
 - {esqlabsR}` now requires `{ospsuite.utils}` version \>= 1.7.0.
+
 - The Importer configuration provided with the template project has been
   updated to include `Gender` in the naming pattern. The new naming pattern is as follows:
   
@@ -47,7 +83,9 @@ did not match the unit (\#823).
   modifying the projectConfiguration object directly, the package will look for
   matching environment variables and build the paths accordingly. A message is
   shown to the user to make this transparent.
+  
 - Complete `sensitivitySpiderPlot` documentation (\#799)
+
 - `parameterPaths` in `sensitivityCalculation()` can now be a named vector. 
 The names will be stored and used as custom labels in all relevant plotting functions (\#811).
 
@@ -55,9 +93,12 @@ The names will be stored and used as custom labels in all relevant plotting func
 ## Minor improvements and bug fixes
 
 - Improved print outputs for all classes
+
 - Classes do not inherit from the deprecated `Printable` class from the `{ospsuite.utils}` package.
+
 - Print methods for all classes are now implemented using the `ospPrint\*` functions 
 introduced in version 1.7.0. of the `{ospsuite.utils}` package.
+
 - Fix when `createPlotsFromExcel` or `createDataCombinedFromExcel` would return 
 wrong names of DataCombined for which the output path for a simulation scenario is not defined (\#800).